sessions/session-05-multi-agents/live-demo/mentorship_team/requirements.txt
# Mentorship Team Dependencies
google-adk>=0.1.0
beautifulsoup4>=4.12.0
requests>=2.31.0
